Changed 'version' column from TINYINT to SMALLINT  - Tables are 'alf_prop_root' and 'alf_audit_app'  - Added property test to update past the SMALLINT rollover to test reversion to 0  - To modify your tables (not absolutely necessary)    alter table alf_audit_app modify version SMALLINT NOT NULL;    alter table alf_prop_root modify version SMALLINT NOT NULL;
git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/HEAD/root@16320 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ config/alfresco/dbscripts/create/3.2/org.hibernate.dialect.MySQLInnoDBDialect/AlfrescoPostCreate-3.2-AuditTables.sql
@@ -20,7 +20,7 @@ CREATE TABLE alf_audit_model
 CREATE TABLE alf_audit_app
 (
    id BIGINT NOT NULL AUTO_INCREMENT,
-   version TINYINT NOT NULL,
+   version SMALLINT NOT NULL,
    app_name_id BIGINT NOT NULL,
    audit_model_id BIGINT NOT NULL,
    disabled_paths_id BIGINT NOT NULL,
